Move CValidationState to consensus subdir
